requirements.txt
streamlit
python-chess
stockfish
torch
numpy
tqdm
